refactor batch repo package functions
